Extracts session output assembly
- Moves transcript-to-line assembly into `crates/agentty/src/ui/session_output_assembly.rs` so `SessionOutput` focuses on layout caching, scrolling, loader effects, and painting.
- Routes cached body and full layout derivation through the assembly module, with focused coverage for queued messages, transcript splitting, prompt rendering, and Tachyon loader status selection.
- Documents the ownership split in `docs/site/content/docs/architecture/module-map.md`.
